Movement Detection
Despite being in the settings since the beginning, alert is moving has not functioned... until now. Atmos Weather now periodically checks your location to see if you are moving, and if so, changes the alerts you receive. The system is designed so that even if you start moving after an alert if first issued, you will still receive the alert.