Radar Bug Fix
I think I actually fixed it this time. The radar bug was cause by animation code running before the radar tiles had actually loaded, resulting in a while loop that ran forever. This is solved by not running the animation before the tiles are loaded. Some timing was also changes to prevent lag.