Another UI Update
Finally tracked down the cause of text running off screen when there wasn't a margin added. Turns out, I had the tab-div class set to the width of the window. (I don't know why I did this, it worked better when width wasn't set.) The issue was fixed and the boilerplate code was removed. Additionally, font size for settings is now smaller for mobile devices to make them less cramped.