Refine welcome window transitions and privacy text
Shortened animation durations for smoother transitions in the welcome window.  Replaced <center> tags with a 'centered' CSS class and added appropriate alt attributes for images to improve accessibility. Removed specific services from the intro privacy text so that it does not have to be updated when services are added (the privacy statement will still be updated.)